replaced main page with the index.html
