+++ frontend/pages/index.tsx
@@ -1,18 +1,51 @@
+import { Box, Flex, Text, VStack } from "@chakra-ui/react";
+import Navbar from "../components/shared/Navbar";
+import PlayerBar from "../components/shared/PlayerBar";
+import ForYouSection from "../components/home/ForYouSection";
+import ContinueListeningSection from "../components/home/ContinueListening";
+import ExploreGenresSection from "../components/home/ExploreGenres";
+import { podcasts, episodes } from "../utilities/SampleData";
 
-// pages/index.tsx
-import { Box, Button, Center, Text } from "@chakra-ui/react";
-import Navbar from "../components/shared/Navbar"; 
-import AuthHelper from "../helpers/AuthHelper";
-
-export default function Home() {
+const Main = () => {
   return (
     <>
+      {/* Navbar */}
       <Navbar />
-      <Center h="100vh">
-        <Box>
-          <Box p={4}>Main Content Here</Box>
+      <ExploreGenresSection />
+      <Box display="flex" flexDirection="column" px={["1em", "2em", "4em"]} > {/* Adjusted padding */}
+      {/* Main Content */}
+      <Flex flex="1" overflow="hidden">
+        {/* For You Section */}
+        <Box flex="1" borderRadius="35px" backdropFilter="blur(35px)" boxShadow="xl">
+          <Text fontSize="xl" fontWeight="bold" ml={3} mt={3}>
+            For You
+          </Text>
+          <Box overflowY="auto" overflowX="hidden" className="no-scrollbar" height="calc(100vh - 60px - 80px)">
+            <ForYouSection episodes={episodes} />
+          </Box>
         </Box>
-      </Center>
+
+        {/* Right Section */}
+        <Flex flex="1" direction="column" overflow="hidden" height="100%">
+          {/* Continue Listening Section */}
+          <VStack spacing={4} h="50%" borderRadius="35px" backdropFilter="blur(35px)">
+            <Text fontSize="xl" fontWeight="bold" ml={3} mt={3}>
+              Continue Listening
+            </Text>
+            <Box overflowY="auto" overflowX="hidden" className="no-scrollbar" flex="1">
+              <ContinueListeningSection />
+            </Box>
+          </VStack>
+
+        
+        </Flex>
+      </Flex>
+
+      {/* Player Bar */}
+      <PlayerBar {...episodes[0]} />
+    </Box>
     </>
   );
-}
+};
+
+export default Main;
